#2 - improve language handling code
After discussion, we decided it's best to concentrate on English specifically, because we are unsure of the results of an AI that uses references from multiple languages.  We made a list of all English dialects, in order of preference.  I.e., US English first, followed by others most similar to that dialect.  `YouTubeTranscriptApi` doesn't provide a case-insensitive search for language codes, so we
added code to do that.